[#3] 사진 인식 요청, 응답 DTO 정리

+++ src/main/java/com/example/demo/dto/Recognizer/FastAPI/ResponseFoodRecogAPIDto.java
@@ -20,6 +20,6 @@ public class ResponseFoodRecogAPIDto {
 
     private List<String> foodNameList = new ArrayList<>();
 
-    private List<Integer> foodCodeList = new ArrayList<>();
+    private List<Long> foodCodeList = new ArrayList<>();
 
 }
